% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convertCharColsToNumeric}
\alias{convertCharColsToNumeric}
\title{Convert Character Columns with Comma as Decimal Separator to Numeric}
\usage{
convertCharColsToNumeric(x, cols)
}
\arguments{
\item{x}{A data.frame in which the columns will be converted.}

\item{cols}{A vector of column names or indices specifying which columns to convert to numeric.
These columns must be character type.}
}
\value{
A data.frame with the specified character columns converted to numeric.
}
\description{
This function converts specified character columns in a data frame to numeric,
replacing commas with dots as the decimal separator.
The function ensures that the input is a data frame and that the specified columns exist.
}
\examples{
df <- data.frame(A = c("1,23", "2,34", "3,45"), B = c("10,5", "20,6", "30,7"), stringsAsFactors = FALSE)
convertCharColsToNumeric(df, cols = c("A", "B"))
}
\author{
Jemay SALOMON
}
