% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getInfo}
\alias{getInfo}
\title{Retrieve a Value from a DataFrame}
\usage{
getInfo(data, x, colName)
}
\arguments{
\item{data}{DataFrame. The DataFrame containing the data.}

\item{x}{Mixed. The row identifier, which can be numeric (row index) or any value that uniquely identifies a row (e.g., row name).}

\item{colName}{Character. The name of the column from which to retrieve the value. Must be a valid column name in the DataFrame.}
}
\value{
The value at the intersection of the specified row and column in the DataFrame.
}
\description{
\code{getInfo} extracts a value from a specified row and column in a DataFrame.
}
\details{
This function is useful for retrieving specific elements from a DataFrame when you know the
row identifier (which can be numeric, character, or any other type) and the column name.

This function performs the following checks:
\itemize{
\item It checks if \code{colName} exists in the DataFrame columns.
\item If \code{x} is numeric, it checks if it is a valid row index.
\item If \code{x} is not numeric, it checks if it matches a row name or identifier.
}
If any of these conditions are not met, the function will stop and throw an error message.
}
\examples{
# Example usage with numeric row index
data <- data.frame(x = c(1, 2, 3), "Isolate_15FG033" = c(1, 0, 0), "Isolate_15FG038" = c(0, 2, 1))
value <- getInfo(data, 2, "Isolate_15FG033")
print(value)

}
\author{
Jemay SALOMON
}
