% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_GenAbel.R
\name{GWAS_GenABEL}
\alias{GWAS_GenABEL}
\title{GWAS using GenAbel functions}
\usage{
GWAS_GenABEL(gwaa, trait = 1, model = "kinship", nPC = NULL, K_i = TRUE,
  verbose = FALSE)
}
\arguments{
\item{gwaa}{\code{gwaa.data} object.}

\item{trait}{\code{Numerical} or \code{character} indicator to specify which
trait of the gp object should be used. Default = 1.}

\item{model}{\code{Character} argument specifying the type of model that is
used to perform the GWAS analysis: 1) 'null' for a simple model with only the
QTL position; 2) 'PC' model with genetic background correction using the
\code{nPC} first principal components; 3) 'kinship', model with kinship
genetic background correction. Default = 'kinship'.}

\item{nPC}{Optional matrix of principal components for genetic background
correction. Default = NULL.}

\item{K_i}{\code{Logical} specifying if the kinship correction should be done
by removing the markers of the scanned chromosome. Default = TRUE.}

\item{verbose}{\code{Logical} indicating if function outputs should be printed.
Default = FALSE.}
}
\value{
Return:

\item{G_res}{Object of class \code{G_res} representing a data.frame with four
columns: marker identifier, chromosome, position in cM or bp and
-log10(p-value).}
}
\description{
Perform a SNP genome wide association study using optional principal
components terms or a kinship matrix to correct for the genetic background.
The function is a wrapper for the \code{GenAbel} package functions
(Covarrubias-Pazaran, 2016).
}
\details{
The model can be fitted using the kinship containing all markers or removing
the markers of the scanned chromosome (\code{K_i = TRUE}).
}
\examples{

Come later

}
\references{
Aulchenko, Y. S., Ripke, S., Isaacs, A., & Van Duijn, C. M. (2007). GenABEL:
an R library for genome-wide association analysis. Bioinformatics, 23(10),
1294-1296.

Aulchenko, Y. S., De Koning, D. J., & Haley, C. (2007). Genomewide rapid
association using mixed model and regression: a fast and simple method for
genomewide pedigree-based quantitative trait loci association analysis.
Genetics, 177(1), 577-585.
}
\author{
Vincent Garin
}
