% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{gwasRun}
\alias{gwasRun}
\title{Run GWAS using MM4LMM, mlmm, eagle Methods}
\usage{
gwasRun(Y, listXs, listVCov, maxsteps = 6, method = NULL, verbose = TRUE)
}
\arguments{
\item{Y}{A numeric vector representing the phenotype data.}

\item{listXs}{A list containing the genotype data matrix, with the key \code{"X"} representing the matrix of genotypes.}

\item{listVCov}{A list containing variance-covariance matrices. The key \code{"K"} represents the kinship matrix, and optionally, the key \code{"Error"} can represent an error variance matrix.}

\item{maxsteps}{An optional numeric value specifying the maximum number of steps for the \code{mlmm} method. This argument is required if the \code{method} is \code{"mlmm"}.}

\item{method}{A character string specifying the method to use for the GWAS. Options are \code{"mm4lmm"}, \code{"mlmm"}, or \code{"eagle"}.}

\item{verbose}{A logical value indicating whether to print progress messages. Default is \code{TRUE}.}
}
\value{
A list containing the results of the GWAS. The structure of the list varies depending on the method used:
\itemize{
\item For \code{"mm4lmm"}: The list includes the estimated model (\code{mmest}), test results (\code{out.test}), and p-values (\code{res.mm4lmm}).
\item For \code{"mlmm"}: The list includes the results of the \code{mlmm} function (\code{mygwas}).
\item For \code{"eagle"}: The list includes the results of the \code{eagle}
}
}
\description{
The \code{gwasRun} function performs a Genome-Wide Association Study (GWAS) using one of three specified methods: MM4LMM, mlmm, or eagle. The function takes phenotype data, genotype data, and variance-covariance matrices as input and returns a list containing the results of the analysis.
}
\details{
\itemize{
\item \strong{MM4LMM}: Utilizes the \code{MM4LMM} package to fit a linear mixed model and compute p-values.
\item \strong{MLMM}: Uses the \code{mlmm} package to perform a multi-locus mixed-model analysis.
\item \strong{eagle}: Intended for use with the \code{eagle} package to perform GWAS
}
}
\examples{
\dontrun{
  # Example usage for the mm4lmm method
  result <- gwasRun(Y = phenotype_data, listXs = genotype_list, listVCov = covariance_list, method = "mm4lmm")

  # Example usage for the mlmm method
  result <- gwasRun(Y = phenotype_data, listXs = genotype_list, listVCov = covariance_list, maxsteps = 6, method = "mlmm")
  
  # Example usage for the eagle method
  
}

}
\author{
Jemay SALOMON
}
