% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{packageCheck}
\alias{packageCheck}
\title{Check if a Package is Installed}
\usage{
packageCheck(pkg)
}
\arguments{
\item{pkg}{A character string specifying the name of the package to check.}
}
\value{
This function does not return a value. It is used for its side effect of checking for package availability. If the package is not found, the function stops execution with an error.
}
\description{
The \code{packageCheck} function verifies whether a specified package is installed and loaded into R. If the package is not available, the function stops execution and returns an error message indicating that the package is required.
}
\examples{
\dontrun{
  packageCheck("ggplot2")  # Checks if the ggplot2 package is installed
}

}
