% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{QC}
\alias{QC}
\title{Quality control}
\usage{
QC(geno, pheno, MkMiss = 0.1, GenMiss = 0.1, MAFlim = 0.05,
  pheno_imp = "mean", geno_imp = "random")
}
\arguments{
\item{geno}{Genotype marker score matrix in HapMap format.}

\item{pheno}{Data.frame with the first column being character genotype
identifier and the next column being numeric phenotypic values.}

\item{MAFlim}{Critical minor allele frequency under which markers are removed.
Default = 0.05.}

\item{pheno_imp}{Method for imputation of the missing phenotypic values.
One of : 'no', 'random', 'mean'. Default = 'mean'.}

\item{geno_imp}{Method for imputation of the missing marker scores values.
For the moment only 'random'.}

\item{Mkmiss}{Maximum percentage of phenotypic values in the markers.
Default = 0.1.

#' @param Genmiss Maximum percentage of phenotypic values in the genotypes.
Default = 0.1.}
}
\value{
Return:

\code{List} with two objects

\item{gp}{a gpData object containing all data processed data (genotype, map,
phenotype) that can directly be used in the function GWAS_scan()}

\item{genoHapMap}{The sorted genotype marker matrix in HapMap format
that can be used for GWAS analysis in TASSEL using the function GWAS_TASSEL().}
}
\description{
Function performing a quality control on...
}
