% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{space}
\alias{space}
\title{Print Space (Interline)}
\usage{
space(x)
}
\arguments{
\item{x}{A numeric value representing the number of interlines to print. It should be an integer (not decimal).}
}
\value{
Prints the specified number of newline characters.
}
\description{
This function prints a specified number of newline characters (interlines) to create space in the console output.
}
\note{
The function ensures that the input is a non-decimal numeric value.
}
\examples{
space(2)
}
\author{
Jemay SALOMON
}
